requirements.txt
feedparser
trafilatura
faiss-cpu
sentence-transformers
transformers
torch
streamlit
requests
